Move cleanDir and fix memory issue (maybe?)
